Fixed syntax error: the zRotation should be a CGFloat

+++ JumpingBall/JumpingBall/PlayScene.swift
@@ -56,7 +56,7 @@ class PlayScene: SKScene{
         // Rotation Degree
         var rotationDegree = CDouble(self.groundSpeed) * M_PI / 180
         // Rotate the hero
-        self.hero.zRotation -= rotationDegree
+        self.hero.zRotation -= CGFloat(rotationDegree)
         
         // Change the X-Position of the running bar
         self.runningBar.position.x -= CGFloat(self.groundSpeed)
